+++ src/utils/first-sync.ts
@@ -1,11 +1,11 @@
 // Return the first non-null or -undefined result from an array of
 // sync functions
-export default function firstSync<T>(
+export default function firstSync<T, Context>(
 	candidates: ((...args: any[]) => T | void)[]
 ): (...args: any[]) => T | void {
-	return function(...args: any[]) {
-		return candidates.reduce<T | void>((result, candidate) => {
-			return result != null ? result : candidate(...args);
+	return function(this: Context, ...args: any[]) {
+		return candidates.reduce((result, candidate) => {
+			return result != null ? result : candidate.call(this, ...args);
 		}, null);
 	};
 }