+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.0
 *unreleased*
 * Handle known undefined properties when evaluating conditionals ([#2265](https://github.com/rollup/rollup/pull/2265))
+* Access watch events via the plugin context ([#2261](https://github.com/rollup/rollup/pull/2261))
 
 ## 0.60.7
 *2018-06-14*